# Remove m*n locking from resource evaluation and O(m * n * (m+n)) iteration