@apparentlymart 2025-05-22
FWIW that `BenchmarkManyResourceInstances` benchmark used in the description was already present before this change and is already in `package tofu`. It came from https://github.com/opentofu/opentofu/pull/2496, and this was exactly the sort of improvement I was hoping to measure with it someday. 😀